draft move first salary to legacy module
